+++ Changes.md
@@ -1,3 +1,7 @@
+HEAD
+-----------
+- Add lock asserts for Node Manager.
+
 0.9.2
 -----------
 - Improved exception handling in NodeWatcher.